NXP-18968: Add a warning message when user wants to add an existing parameter in template rendering configuration
