+++ nuxeo-template-rendering-api/src/main/java/org/nuxeo/template/api/adapters/TemplateSourceDocument.java
@@ -52,6 +52,11 @@ public interface TemplateSourceDocument {
     public List<TemplateInput> addInput(TemplateInput input);
 
     /**
+     * Returns whether or not the {@link TemplateInput} already exists, based on the name, in the template.
+     */
+    boolean isInputExist(String inputName);
+
+    /**
      * Return the template Type (i.e. the associated {@link TemplateProcessor} name.
      *
      * @return {@link TemplateProcessor} name if any, null otherwise
