+++ nuxeo-template-rendering-core/src/main/java/org/nuxeo/template/adapters/source/TemplateSourceDocumentAdapterImpl.java
@@ -163,6 +163,12 @@ public class TemplateSourceDocumentAdapterImpl extends AbstractTemplateDocument 
         return params;
     }
 
+    @Override
+    public boolean isInputExist(String inputName) {
+        List<TemplateInput> params = getParams();
+        return params != null && params.stream().map(TemplateInput::getName).anyMatch(inputName::equals);
+    }
+
     public String getTemplateType() {
         String ttype = (String) getAdaptedDoc().getPropertyValue(TEMPLATE_TYPE_PROP);
         if (TEMPLATE_TYPE_AUTO.equals(ttype)) {
