+++ nuxeo-template-rendering-web/src/main/resources/OSGI-INF/l10n/messages_en_US.properties
@@ -24,6 +24,7 @@ label.template.parameter.content.source=Source
 label.template.parameter.htmlPreview=Html preview
 label.template.parameter.blobContent=Binary
 label.template.parameter.loop=Loop
+label.template.parameter.already.exist=Parameter '{0}' already exists
 
 label.template.canceleditparameters = Cancel
 label.template.editparameters = Modify
